+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/strategies/base.py
@@ -1,7 +1,8 @@
 import asyncio
 import random
 from abc import ABC, abstractmethod
-from collections.abc import Awaitable, Callable, Iterable
+from collections import defaultdict
+from collections.abc import Awaitable, Callable, Iterable, Sequence
 from typing import ClassVar, ParamSpec, TypeVar
 
 from pydantic import BaseModel, Field
@@ -9,29 +10,17 @@ from pydantic import BaseModel, Field
 from ragbits.core.utils.config_handling import WithConstructionConfig
 from ragbits.core.vector_stores.base import VectorStore
 from ragbits.document_search.documents.document import Document, DocumentMeta
-from ragbits.document_search.documents.element import Element
+from ragbits.document_search.documents.element import Element, IntermediateElement
 from ragbits.document_search.documents.sources import Source
 from ragbits.document_search.ingestion import strategies
 from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
-from ragbits.document_search.ingestion.providers.base import BaseProvider
+from ragbits.document_search.ingestion.intermediate_handlers.base import BaseIntermediateHandler
 
 _CallP = ParamSpec("_CallP")
 _CallReturnT = TypeVar("_CallReturnT")
 
 
-class IngestTaskResult(BaseModel):
-    """
-    Represents the result of the document ingest tast.
-    """
-
-    class Config:  # noqa: D106
-        arbitrary_types_allowed = True
-
-    document_uri: str
-    response: list[Element] | BaseException
-
-
-class IngestSummaryResult(BaseModel):
+class IngestDocumentResult(BaseModel):
     """
     Represents the result of the document ingest execution.
     """
@@ -49,8 +38,8 @@ class IngestExecutionResult(BaseModel):
     Represents the result of the documents ingest execution.
     """
 
-    successful: list[IngestSummaryResult] = Field(default_factory=list)
-    failed: list[IngestSummaryResult] = Field(default_factory=list)
+    successful: list[IngestDocumentResult] = Field(default_factory=list)
+    failed: list[IngestDocumentResult] = Field(default_factory=list)
 
 
 class IngestStrategy(WithConstructionConfig, ABC):
@@ -78,8 +67,8 @@ class IngestStrategy(WithConstructionConfig, ABC):
         self,
         documents: Iterable[DocumentMeta | Document | Source],
         vector_store: VectorStore,
-        processor_router: DocumentProcessorRouter,
-        processor_overwrite: BaseProvider | None = None,
+        parser_router: DocumentProcessorRouter,
+        enricher_router: dict[type[IntermediateElement], BaseIntermediateHandler],
     ) -> IngestExecutionResult:
         """
         Ingest documents.
@@ -87,8 +76,8 @@ class IngestStrategy(WithConstructionConfig, ABC):
         Args:
             documents: The documents to ingest.
             vector_store: The vector store to store document chunks.
-            processor_router: The document processor router to use.
-            processor_overwrite: Forces the use of a specific processor, instead of the one provided by the router.
+            parser_router: The document parser router to use.
+            enricher_router: The intermediate element enricher router to use.
 
         Returns:
             The ingest execution result.
@@ -97,34 +86,30 @@ class IngestStrategy(WithConstructionConfig, ABC):
     async def _call_with_error_handling(
         self,
         executable: Callable[_CallP, Awaitable[_CallReturnT]],
-        return_exception: bool = False,
         *executable_args: _CallP.args,
         **executable_kwargs: _CallP.kwargs,
-    ) -> _CallReturnT | BaseException:
+    ) -> _CallReturnT:
         """
         Call executable with standarized error handling.
         If an error occurs, the executable is retried `num_retries` times using randomized exponential backoff.
 
         Args:
             executable: The callable function to execute.
-            return_exception: If True, return the exception instead of raising it after all retries.
             executable_args: Positional arguments to pass to the executable.
             executable_kwargs: Keyword arguments to pass to the executable.
 
         Returns:
             The result of the executable if successful.
-            If all retries fail and `return_exception` is True, returns the last encountered exception.
-            Otherwise, raises the exception after all retries are exhausted.
+
+        Raises:
+            Exception: The last encountered exception after all retries are exhausted.
         """
         for i in range(max(0, self.num_retries) + 1):
             try:
                 return await executable(*executable_args, **executable_kwargs)
             except Exception as exc:
                 if i == self.num_retries:
-                    if return_exception:
-                        return exc
-                    else:
-                        raise exc
+                    raise exc
 
                 delay = min(2**i * self.backoff_multiplier, self.backoff_max)
                 delay = random.uniform(0, delay) if delay < self.backoff_max else random.uniform(0, self.backoff_max)  # noqa S311
@@ -135,16 +120,14 @@ class IngestStrategy(WithConstructionConfig, ABC):
     @staticmethod
     async def _parse_document(
         document: DocumentMeta | Document | Source,
-        processor_router: DocumentProcessorRouter,
-        processor_overwrite: BaseProvider | None = None,
-    ) -> list[Element]:
+        parser_router: DocumentProcessorRouter,
+    ) -> Sequence[Element | IntermediateElement]:
         """
         Parse a single document and return the elements.
 
         Args:
             document: The document to parse.
-            processor_router: The document processor router to use.
-            processor_overwrite: Forces the use of a specific processor, instead of the one provided by the router.
+            parser_router: The document parser router to use.
 
         Returns:
             The list of elements.
@@ -156,11 +139,40 @@ class IngestStrategy(WithConstructionConfig, ABC):
             if isinstance(document, DocumentMeta)
             else document.metadata
         )
-        processor = processor_overwrite or processor_router.get_provider(document_meta)
-        return await processor.process(document_meta)
+        parser = parser_router.get_provider(document_meta)
+        return await parser.process(document_meta)
 
     @staticmethod
-    async def _remove_elements(elements: list[Element], vector_store: VectorStore) -> None:
+    async def _enrich_elements(
+        elements: Iterable[IntermediateElement],
+        enricher_router: dict[type[IntermediateElement], BaseIntermediateHandler],
+    ) -> list[Element]:
+        """
+        Enrich intermediate elements.
+
+        Args:
+            elements: The document elements to enrich.
+            enricher_router: The intermediate element enricher router to use.
+
+        Returns:
+            The list of enriched elements.
+        """
+        grouped_intermediate_elements: dict[type, list[IntermediateElement]] = defaultdict(list)
+        for element in elements:
+            if isinstance(element, IntermediateElement):
+                grouped_intermediate_elements[type(element)].append(element)
+
+        grouped_enriched_elements = await asyncio.gather(
+            *[
+                enricher.process(intermediate_elements)
+                for element_type, intermediate_elements in grouped_intermediate_elements.items()
+                if (enricher := enricher_router.get(element_type))
+            ]
+        )
+        return [element for enriched_elements in grouped_enriched_elements for element in enriched_elements]
+
+    @staticmethod
+    async def _remove_elements(elements: Iterable[Element], vector_store: VectorStore) -> None:
         """
         Remove entries from the vector store whose source id is present in the elements metadata.
 
@@ -179,9 +191,9 @@ class IngestStrategy(WithConstructionConfig, ABC):
             await vector_store.remove(ids_to_delete)
 
     @staticmethod
-    async def _insert_elements(elements: list[Element], vector_store: VectorStore) -> None:
+    async def _insert_elements(elements: Iterable[Element], vector_store: VectorStore) -> None:
         """
-        Insert Elements into the vector store.
+        Insert elements into the vector store.
 
         Args:
             elements: The list of elements to insert.